validation errors separated to own util class
